+++ services/bet-service/src/main/java/com/beetclick/betservice/event/BetPlaced.java
@@ -1,5 +1,6 @@
 package com.beetclick.betservice.event;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record BetPlaced(UUID userId, UUID matchId, double amount) {}
+public record BetPlaced(UUID userId, UUID matchId, BigDecimal amount) {}
