+++ services/bet-service/src/main/java/com/beetclick/betservice/event/WalletResponse.java
@@ -1,5 +1,6 @@
 package com.beetclick.betservice.event;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record WalletResponse (UUID betId, UUID userId, double balance) {}
+public record WalletResponse (UUID betId, UUID userId, BigDecimal balance) {}
